Add option for sourcepath in the config file
Previously the checker passed only the current buffer as input to the
kotlin compiler, everything else was passed as classpath. This meant
an access to an internal program element (function, type, variable, etc)
was reported as error. This commit adds a sourcepath option that will be
passed to the compiler during a check.