Improve matching of some soft keywords
Words like "value" and "internal" are very common in code, yet they are
only parsed as modifiers if they're present before a class or some other
declaration. Use positive lookahead with the end-of-match token "\ze" to
only highlight them in this case too. This will prevent them from being
highlighted in multi-line case, but it's not very common to have a
declaration and its modifiers on separate lines anyway.